+++ src/embed_tests/NeedsReinit/TestPyInitialize.cs
@@ -2,9 +2,10 @@ using System;
 using NUnit.Framework;
 using Python.Runtime;
 
-namespace Python.EmbeddingTest
+namespace Python.EmbeddingTest.NeedsReinit
 {
-    public class PyInitializeTest
+    [Category("NeedsReinit")]
+    public class TestPyInitialize : StopAndRestartEngine
     {
         /// <summary>
         /// Tests issue with multiple simple Initialize/Shutdowns.
@@ -42,8 +43,8 @@ namespace Python.EmbeddingTest
                 {
                     using var v0 = argv[0];
                     using var v1 = argv[1];
-                    Assert.AreEqual(args[0], v0.ToString());
-                    Assert.AreEqual(args[1], v1.ToString());
+                    Assert.That(v0.ToString(), Is.EqualTo(args[0]));
+                    Assert.That(v1.ToString(), Is.EqualTo(args[1]));
                 }
             }
         }