+++ src/embed_tests/TestGILState.cs
@@ -17,17 +17,5 @@ namespace Python.EmbeddingTest
                     gilState.Dispose();
             }
         }
-
-        [OneTimeSetUp]
-        public void SetUp()
-        {
-            PythonEngine.Initialize();
-        }
-
-        [OneTimeTearDown]
-        public void Dispose()
-        {
-            PythonEngine.Shutdown();
-        }
     }
 }