+++ src/runtime/ClassManager.cs
@@ -213,6 +213,7 @@ namespace Python.Runtime
             ClassInfo info = GetClassInfo(type, impl);
 
             impl.indexer = info.indexer;
+            impl.del = info.del;
             impl.richcompare.Clear();
 
 
@@ -538,6 +539,21 @@ namespace Python.Runtime
 
                 ob = new MethodObject(type, name, mlist);
                 ci.members[name] = ob.AllocObject();
+                if (name == nameof(IDictionary<int, int>.Remove)
+                    && mlist.Any(m => m.DeclaringType?.GetInterfaces()
+                        .Any(i => i.TryGetGenericDefinition() == typeof(IDictionary<,>)) is true))
+                {
+                    ci.del = new();
+                    ci.del.AddRange(mlist.Where(m => !m.IsStatic));
+                }
+                else if (name == nameof(IList<int>.RemoveAt)
+                         && mlist.Any(m => m.DeclaringType?.GetInterfaces()
+                             .Any(i => i.TryGetGenericDefinition() == typeof(IList<>)) is true))
+                {
+                    ci.del = new();
+                    ci.del.AddRange(mlist.Where(m => !m.IsStatic));
+                }
+
                 if (mlist.Any(OperatorMethod.IsOperatorMethod))
                 {
                     string pyName = OperatorMethod.GetPyMethodName(name);
@@ -581,6 +597,7 @@ namespace Python.Runtime
         private class ClassInfo
         {
             public Indexer? indexer;
+            public MethodBinder? del;
             public readonly Dictionary<string, PyObject> members = new();
 
             internal ClassInfo()