+++ CHANGELOG.md
@@ -10,10 +10,12 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 ### Added
 
 - Add context manager protocol for .NET IDisposable types, allowing use of `with` statements for IDisposable objects (#9c73c35)
+-  Support `del obj[...]` for types derived from `IList<T>` and `IDictionary<K, V>`
 
 ### Changed
 ### Fixed
 
+-  Fixed crash when trying to `del clrObj[...]` for non-arrays
 - ci: properly exclude job (#2542)
 
 ## [3.0.5](https://github.com/pythonnet/pythonnet/releases/tag/v3.0.5) - 2024-12-13