+++ CHANGELOG.md
@@ -10,12 +10,13 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 ### Added
 
 -  Support `del obj[...]` for types derived from `IList<T>` and `IDictionary<K, V>`
+-  Support for .NET Framework 4.6.1 (#2701)
 
 ### Changed
 ### Fixed
 
 -  Fixed crash when trying to `del clrObj[...]` for non-arrays
-- ci: properly exclude job (#2542)
+-  ci: properly exclude job (#2542)
 
 ## [3.0.5](https://github.com/pythonnet/pythonnet/releases/tag/v3.0.5) - 2024-12-13